+++ core/app/models/spree/product/scopes.rb
@@ -23,7 +23,7 @@ module Spree
         # We should not define price scopes here, as they require something slightly different
         next if name.to_s.include?("master_price")
         parts = name.to_s.match(/(.*)_by_(.*)/)
-        self.scope(name.to_s, -> { relation.order("#{Product.quoted_table_name}.#{parts[2]} #{parts[1] == 'ascend' ?  "ASC" : "DESC"}") })
+        self.scope(name.to_s, -> { order("#{Product.quoted_table_name}.#{parts[2]} #{parts[1] == 'ascend' ?  "ASC" : "DESC"}") })
       end
     end
 
@@ -67,11 +67,9 @@ module Spree
     #
     #   SELECT COUNT(*) ...
     add_search_scope :in_taxon do |taxon|
-      select("spree_products.id, spree_products.*").
-      where(id: Classification.select('spree_products_taxons.product_id').
-            joins(:taxon).
-            where(Taxon.table_name => { :id => taxon.self_and_descendants.pluck(:id) })
-           )
+      includes(:classifications).
+      where("spree_products_taxons.taxon_id" => taxon.self_and_descendants.pluck(:id)).
+      order("spree_products_taxons.position ASC")
     end
 
     # This scope selects products in all taxons AND all its descendants