+++ core/app/models/spree/variant.rb
@@ -3,10 +3,11 @@ module Spree
     acts_as_paranoid
 
     belongs_to :product, touch: true, class_name: 'Spree::Product'
+    belongs_to :tax_category, class_name: 'Spree::TaxCategory'
 
     delegate_belongs_to :product, :name, :description, :permalink, :available_on,
-                        :tax_category_id, :shipping_category_id, :meta_description,
-                        :meta_keywords, :tax_category, :shipping_category
+                        :shipping_category_id, :meta_description, :meta_keywords,
+                        :shipping_category
 
     has_many :inventory_units
     has_many :line_items
@@ -23,7 +24,7 @@ module Spree
       class_name: 'Spree::Price',
       dependent: :destroy
 
-    delegate_belongs_to :default_price, :display_price, :display_amount, :price, :price=, :currency 
+    delegate_belongs_to :default_price, :display_price, :display_amount, :price, :price=, :currency
 
     has_many :prices,
       class_name: 'Spree::Price',
@@ -44,6 +45,14 @@ module Spree
 
     def self.active(currency = nil)
       joins(:prices).where(deleted_at: nil).where('spree_prices.currency' => currency || Spree::Config[:currency]).where('spree_prices.amount IS NOT NULL')
+    end
+
+    def tax_category
+      if self[:tax_category_id].nil?
+        product.tax_category
+      else
+        TaxCategory.find(self[:tax_category_id])
+      end
     end
 
     def cost_price=(price)
@@ -127,19 +136,29 @@ module Spree
       "#{name} - #{sku}"
     end
 
+    def sku_and_options_text
+      "#{sku} #{options_text}".strip
+    end
+
     # Product may be created with deleted_at already set,
     # which would make AR's default finder return nil.
     # This is a stopgap for that little problem.
     def product
       Spree::Product.unscoped { super }
     end
-    
+
     def in_stock?(quantity=1)
       Spree::Stock::Quantifier.new(self).can_supply?(quantity)
     end
 
     def total_on_hand
       Spree::Stock::Quantifier.new(self).total_on_hand
+    end
+
+    # Shortcut method to determine if inventory tracking is enabled for this variant
+    # This considers both variant tracking flag and site-wide inventory tracking settings
+    def should_track_inventory?
+      self.track_inventory? && Spree::Config.track_inventory_levels
     end
 
     private